flamenco, blockstore: deserialize per-batch and allow trailing bytes
Agave has this notion of a batch of data shreds.  The presence of one
of the COMPLETE flags in a data shred marks the end of a batch.
Moreover, bincode deserialization of microblocks is performed on a
per-batch basis.  Precisely a single array of microblocks is expected
to be deserialized from a batch.  Trailing bytes in each batch are
ignored by default.

Block parsers in both the runtime and the blockstore lack the notion
of a batch, and consequently do not properly handle blocks with
trailing bytes, such as
https://explorer.solana.com/block/308445708
which has 7335 trailing bytes in the last batch.